Adding the link check tool to travis
Going ahead with https://github.com/bmuschko/link-verifier as the choice
of link checking tool. It is written in Go, is quite fast and accurate.

Currently keeping the tool check not to cause the build failure upon
finding broken links. Build failure will be kept in enforcing mode once
the sanity of document links is in good health.

Signed-off-by: Jayesh Kumar <k8s.cncf@gmail.com>